// src/types/technology/Technology.ts
export type Technology =
  | 'React'
  | 'NodeJs'
  | 'Spring'
  | 'ReactNative'
  | 'Electron'
  | 'NextJs'
  | 'Expo'
  | 'NestJs'
  | 'Express'
  | 'Fastify'
  | 'JavaScript'
  | 'TypeScript'
  | 'Java'
  | 'C'
  | 'C++'
  | 'Python'
  | 'Docker'
  | 'Android'
  | 'Linux'
  | 'Arch'
  | 'Ubuntu'
  | 'Swagger'
  | 'Jest'
  | 'AWS'
  | 'Firebase'
  | 'StyledComponents'
  | 'TailwindCSS'
  | 'TypeORM'
  | 'Prisma'
  | 'PostgreSQL'
  | 'MongoDB'
  | 'Redis'
  | 'SQLite'
  | 'HTML'
  | 'CSS';